All features working. Added README to give information about the code

+++ feature_extraction/services/time_service.py
@@ -6,5 +6,5 @@ import numpy as np
 def calculate_time_features(entry):
     dt = datetime.datetime.strptime(entry["postTimestamp"], '%a %b %d %H:%M:%S %z %Y')
 
-    hours = np.floor(dt.timestamp() / 3600)
+    hours = int(np.floor(dt.timestamp() / 3600))
     return hours
